Gracefully exit if env variables are not set (#10496)

+++ sdk/search/search-documents/samples/javascript/src/dataSourceConnections/deleteDataSourceConnectionByName.js
@@ -12,10 +12,15 @@ const apiKey = process.env.SEARCH_API_KEY || "";
 
 async function main() {
   console.log(`Running Delete Datasource Connection Sample....`);
-
+  if (!endpoint || !apiKey) {
+    console.log("Make sure to set valid values for endpoint and apiKey with proper authorization.");
+    return;
+  }
   const client = new SearchIndexerClient(endpoint, new AzureKeyCredential(apiKey));
   console.log(`Deleting Datasource Connection my-data-source-2`);
   await client.deleteDataSourceConnection("my-data-source-2")
 }
 
-main();
+main().catch((err) => {
+  console.error("The sample encountered an error:", err);
+});
